Add files via upload
Takes in text file. Priority queue sorts words in fashion where words that start with vowels have the highest priority, while words that start with invalid characters have the lowest. 

The priority list is then emptied into a binary tree where the inorder traversal lists the words in order of increasing length regardless if they start with a vowel or not.